Update LICENSE
Modified the MIT license with the intent that anyone who profits from this solution should make donation to the Icelandic Childrens hospital.

Note: this only applies to those who profit through sub-licensing, licensing or sell services using this solutions. An employee of a company is allowed to use this software as stated in the license without the requirement to make a donation. If the Software is beneficial for this individual/company I, the author, would appreciate if you would make this donation nevertheless.